refactor(news/details): FSD 구조에 알맞게 컴포넌트 이동
