+++ src/app/[locale]/news/[...source]/__container.tsx
@@ -4,7 +4,13 @@ import { Button, buttonVariants } from '@/shared/components/ui/button';
 import { Link } from '@/shared/i18n/routing';
 import { SquareArrowOutUpRight, Undo2 } from 'lucide-react';
 import sanitizeHtml from 'sanitize-html';
-import { NewsImage, useNewsBySource, useScrapedContent, useSummary } from '@/features/news';
+import {
+  NewsHeader,
+  NewsImage,
+  useNewsBySource,
+  useScrapedContent,
+  useSummary
+} from '@/features/news';
 import { useParams } from 'next/navigation';
 import { unstable_ViewTransition as ViewTransition } from 'react';
 import { useTranslations } from 'next-intl';
@@ -61,24 +67,3 @@ export default function NewsIdContainer() {
     </article>
   );
 }
-
-const NewsHeader = ({
-  title,
-  source,
-  pubDate
-}: {
-  title: string;
-  source: string;
-  pubDate: string;
-}) => (
-  <header className="mb-[10px] inline-block">
-    <h1 className="text-lg font-bold">{title}</h1>
-    <div className="flex items-center text-xs md:text-sm text-gray-500 dark:text-gray-400">
-      <span>{source}</span>
-      <span className="mx-1.5 sm:mx-2">•</span>
-      <span>{pubDate}</span>
-    </div>
-  </header>
-);
-
-NewsHeader.displayName = 'NewsHeader';
