+++ src/features/news/ui/index.tsx
@@ -1,3 +1,4 @@
+export { default as NewsHeader } from './NewsHeader';
 export { default as NewsImage } from './NewsImage';
 export { default as NewsItem } from './NewsItem';
 export { default as NewsItemMeta } from './NewsItemMeta';
